Reproducible Analysis
I changed the coding for pacman so the document could be knitted. I added more detail into the read.me file so a person knows the steps to take to be able to reproduce this analysis. I also made some changes to my cleaning of variables as well as the format of the HTML document.